Clean up helper function, move to own file for reusability
